+++ core/persistence/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/persistence/data/MultiCommunityModel.kt
@@ -3,6 +3,6 @@ package com.github.diegoberaldin.raccoonforlemmy.core.persistence.data
 data class MultiCommunityModel(
     val id: Long? = null,
     val name: String = "",
-    val communityIds: List<Int> = emptyList(),
+    val communityIds: List<Long> = emptyList(),
     val icon: String? = null,
 )
