+++ core/utils/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/utils/cache/DefaultLruCache.kt
@@ -5,11 +5,11 @@ import kotlinx.coroutines.sync.withLock
 
 internal class DefaultLruCache<T>(val size: Int) : LruCache<T> {
 
-    private val values: MutableMap<Int, T> = mutableMapOf()
-    private var lastUsedIds: List<Int> = listOf()
+    private val values: MutableMap<Long, T> = mutableMapOf()
+    private var lastUsedIds: List<Long> = listOf()
     private val mutex = Mutex()
 
-    override suspend fun get(key: Int): T? = mutex.withLock {
+    override suspend fun get(key: Long): T? = mutex.withLock {
         val res = values[key]
         if (res != null) {
             moveAtTheBeginning(key)
@@ -17,7 +17,7 @@ internal class DefaultLruCache<T>(val size: Int) : LruCache<T> {
         return res
     }
 
-    override suspend fun put(value: T, key: Int) = mutex.withLock {
+    override suspend fun put(value: T, key: Long) = mutex.withLock {
         val old = values[key]
         if (old != null) {
             // already existing element
@@ -43,7 +43,7 @@ internal class DefaultLruCache<T>(val size: Int) : LruCache<T> {
         }
     }
 
-    private fun moveAtTheBeginning(key: Int) {
+    private fun moveAtTheBeginning(key: Long) {
         lastUsedIds = buildList {
             this += key
             this += (lastUsedIds - key)
