+++ domain/lemmy/data/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/domain/lemmy/data/CommentReportModel.kt
@@ -1,10 +1,10 @@
 package com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.data
 
 data class CommentReportModel(
-    val id: Int = 0,
+    val id: Long = 0,
     val creator: UserModel? = null,
-    val commentId: Int = 0,
-    val postId: Int = 0,
+    val commentId: Long = 0,
+    val postId: Long = 0,
     val originalText: String? = null,
     val reason: String? = null,
     val resolved: Boolean = false,
