+++ domain/lemmy/data/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/domain/lemmy/data/ModlogItem.kt
@@ -4,11 +4,11 @@ sealed class ModlogItem(
     val type: ModlogItemType,
 ) {
 
-    abstract val id: Int
+    abstract val id: Long
     abstract val date: String?
 
     data class ModRemovePost(
-        override val id: Int,
+        override val id: Long,
         override val date: String? = null,
         val removed: Boolean = false,
         val user: UserModel? = null,
@@ -18,7 +18,7 @@ sealed class ModlogItem(
     ) : ModlogItem(type = ModlogItemType.ModRemovePost)
 
     data class ModLockPost(
-        override val id: Int,
+        override val id: Long,
         override val date: String? = null,
         val locked: Boolean = false,
         val moderator: UserModel? = null,
@@ -26,7 +26,7 @@ sealed class ModlogItem(
     ) : ModlogItem(type = ModlogItemType.ModLockPost)
 
     data class ModFeaturePost(
-        override val id: Int,
+        override val id: Long,
         override val date: String? = null,
         val featured: Boolean = false,
         val moderator: UserModel? = null,
@@ -35,7 +35,7 @@ sealed class ModlogItem(
         ModlogItem(type = ModlogItemType.ModFeaturePost)
 
     data class ModRemoveComment(
-        override val id: Int,
+        override val id: Long,
         override val date: String? = null,
         val removed: Boolean = false,
         val user: UserModel? = null,
@@ -46,7 +46,7 @@ sealed class ModlogItem(
     ) : ModlogItem(type = ModlogItemType.ModRemoveComment)
 
     data class ModBanFromCommunity(
-        override val id: Int,
+        override val id: Long,
         override val date: String? = null,
         val banned: Boolean = false,
         val user: UserModel? = null,
@@ -54,7 +54,7 @@ sealed class ModlogItem(
     ) : ModlogItem(type = ModlogItemType.ModBanFromCommunity)
 
     data class ModAdd(
-        override val id: Int,
+        override val id: Long,
         override val date: String? = null,
         val removed: Boolean = false,
         val user: UserModel? = null,
@@ -62,7 +62,7 @@ sealed class ModlogItem(
     ) : ModlogItem(type = ModlogItemType.ModAdd)
 
     data class ModAddCommunity(
-        override val id: Int,
+        override val id: Long,
         override val date: String? = null,
         val removed: Boolean = false,
         val user: UserModel? = null,
@@ -70,32 +70,32 @@ sealed class ModlogItem(
     ) : ModlogItem(type = ModlogItemType.ModAddCommunity)
 
     data class AdminPurgeCommunity(
-        override val id: Int,
+        override val id: Long,
         override val date: String? = null,
         val admin: UserModel? = null,
     ) : ModlogItem(type = ModlogItemType.AdminPurgeCommunity)
 
     data class AdminPurgeComment(
-        override val id: Int,
+        override val id: Long,
         override val date: String? = null,
         val post: PostModel? = null,
         val admin: UserModel? = null,
     ) : ModlogItem(type = ModlogItemType.AdminPurgeComment)
 
     data class AdminPurgePerson(
-        override val id: Int,
+        override val id: Long,
         override val date: String? = null,
         val admin: UserModel? = null,
     ) : ModlogItem(type = ModlogItemType.AdminPurgePerson)
 
     data class AdminPurgePost(
-        override val id: Int,
+        override val id: Long,
         override val date: String? = null,
         val admin: UserModel? = null,
     ) : ModlogItem(type = ModlogItemType.AdminPurgePost)
 
     data class ModBan(
-        override val id: Int,
+        override val id: Long,
         override val date: String? = null,
         val banned: Boolean = false,
         val user: UserModel? = null,
@@ -103,7 +103,7 @@ sealed class ModlogItem(
     ) : ModlogItem(type = ModlogItemType.ModBan)
 
     data class HideCommunity(
-        override val id: Int,
+        override val id: Long,
         override val date: String? = null,
         val hidden: Boolean = false,
         val community: CommunityModel? = null,
@@ -111,14 +111,14 @@ sealed class ModlogItem(
     ) : ModlogItem(type = ModlogItemType.ModHideCommunity)
 
     data class RemoveCommunity(
-        override val id: Int,
+        override val id: Long,
         override val date: String? = null,
         val community: CommunityModel? = null,
         val moderator: UserModel? = null,
     ) : ModlogItem(type = ModlogItemType.ModRemoveCommunity)
 
     data class ModTransferCommunity(
-        override val id: Int,
+        override val id: Long,
         override val date: String? = null,
         val user: UserModel? = null,
         val moderator: UserModel? = null,
