+++ domain/lemmy/data/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/domain/lemmy/data/PostReportModel.kt
@@ -3,9 +3,9 @@ package com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.data
 import com.github.diegoberaldin.raccoonforlemmy.core.utils.looksLikeAnImage
 
 data class PostReportModel(
-    val id: Int = 0,
+    val id: Long = 0,
     val creator: UserModel? = null,
-    val postId: Int = 0,
+    val postId: Long = 0,
     val reason: String? = null,
     val originalTitle: String? = null,
     val originalText: String? = null,
