+++ domain/lemmy/repository/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/domain/lemmy/repository/DefaultLemmyItemCache.kt
@@ -20,23 +20,23 @@ internal class DefaultLemmyItemCache(
         postCache.put(value = value, key = value.id)
     }
 
-    override suspend fun getPost(id: Int): PostModel? = withContext(Dispatchers.IO) { postCache.get(id) }
+    override suspend fun getPost(id: Long): PostModel? = withContext(Dispatchers.IO) { postCache.get(id) }
 
     override suspend fun putComment(value: CommentModel) = withContext(Dispatchers.IO) {
         commentCache.put(value = value, key = value.id)
     }
 
-    override suspend fun getComment(id: Int): CommentModel? = withContext(Dispatchers.IO) { commentCache.get(id) }
+    override suspend fun getComment(id: Long): CommentModel? = withContext(Dispatchers.IO) { commentCache.get(id) }
 
     override suspend fun putCommunity(value: CommunityModel) = withContext(Dispatchers.IO) {
         communityCache.put(value = value, key = value.id)
     }
 
-    override suspend fun getCommunity(id: Int): CommunityModel?  = withContext(Dispatchers.IO) { communityCache.get(id) }
+    override suspend fun getCommunity(id: Long): CommunityModel?  = withContext(Dispatchers.IO) { communityCache.get(id) }
 
     override suspend fun putUser(value: UserModel) = withContext(Dispatchers.IO) {
         userCache.put(value = value, key = value.id)
     }
 
-    override suspend fun getUser(id: Int): UserModel? = withContext(Dispatchers.IO) { userCache.get(id) }
+    override suspend fun getUser(id: Long): UserModel? = withContext(Dispatchers.IO) { userCache.get(id) }
 }
